refacture tests a bit; add settings for celery for tests
